move expiration enum to its own submodule
